Refactor package names
As discussed in #55, this is now the packages structure:

* it.contactlab.hub.sdk.java : main entry point of the SDK.
  (Auth, ContactHub, AsyncContactHub)

* it.contactlab.hub.sdk.java.models : public data model.

* it.contactlab.hub.sdk.java.exceptions : SDK exceptions.

* it.contactlab.hub.sdk.java.internal : all implementation details that
  should never be used directly by client code.
