+++ example/src/main/java/com/example/async/Example.java
@@ -1,7 +1,7 @@
 package com.example.async;
 
+import it.contactlab.hub.sdk.java.AsyncContactHub;
 import it.contactlab.hub.sdk.java.Auth;
-import it.contactlab.hub.sdk.java.async.ContactHub;
 import it.contactlab.hub.sdk.java.models.Customer;
 
 /**
@@ -21,7 +21,7 @@ public class Example {
       System.getenv("CONTACTHUB_TEST_NODE_ID")
     );
 
-    final ContactHub ch = new ContactHub(auth);
+    final AsyncContactHub ch = new AsyncContactHub(auth);
 
     System.out.println("-----------------------------------");
     System.out.println("Retrieving customers' phone numbers");
