+++ example/src/main/java/com/example/sync/Example.java
@@ -1,11 +1,11 @@
 package com.example.sync;
 
 import it.contactlab.hub.sdk.java.Auth;
+import it.contactlab.hub.sdk.java.ContactHub;
 import it.contactlab.hub.sdk.java.models.Customer;
 import it.contactlab.hub.sdk.java.models.CustomerTags;
 import it.contactlab.hub.sdk.java.models.base.BaseProperties;
 import it.contactlab.hub.sdk.java.models.base.Contacts;
-import it.contactlab.hub.sdk.java.sync.ContactHub;
 
 import java.security.SecureRandom;
 import java.util.Arrays;
