+++ src/main/java/it/contactlab/hub/sdk/java/internal/api/SessionApi.java
@@ -1,4 +1,4 @@
-package it.contactlab.hub.sdk.java.api;
+package it.contactlab.hub.sdk.java.internal.api;
 
 import it.contactlab.hub.sdk.java.Auth;
 import it.contactlab.hub.sdk.java.exceptions.HttpException;
