+++ src/main/java/it/contactlab/hub/sdk/java/internal/gson/ContactHubGson.java
