+++ src/test/scala/integration/DataGenerators.scala
@@ -1,4 +1,4 @@
-package it.contactlab.hub.sdk.java.sync.test.integration
+package it.contactlab.hub.sdk.java.test.integration
 
 import it.contactlab.hub.sdk.java.models._, base._
